Backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-google-genai
langchain-core
gradio-client
beautifulsoup4
Pillow
